Fix hash join counters.

+++ src/execution/vm/vm.cpp
@@ -1490,7 +1490,7 @@ void VM::Interpret(const uint8_t *ip, Frame *frame) {  // NOLINT
   }
 
   OP(JoinHashTableGetTupleCount) : {
-    auto *result = frame->LocalAt<uint64_t *>(READ_LOCAL_ID());
+    auto *result = frame->LocalAt<uint32_t *>(READ_LOCAL_ID());
     auto *join_hash_table = frame->LocalAt<sql::JoinHashTable *>(READ_LOCAL_ID());
     OpJoinHashTableGetTupleCount(result, join_hash_table);
     DISPATCH_NEXT();
